updated submlime settings

+++ pythonSource/logger_update.py
@@ -1,9 +1,20 @@
 import os
 import sys
 import socket
+import hashlib
 import argparse
 
 import logger
+
+
+def md5_file(file_path):
+    m = hashlib.md5()
+    with open(file_path, 'rb') as file:
+        cts = file.read(1024)
+        while cts:
+            m.update(cts)
+            cts = file.read(1024)
+    return str(m.hexdigest())
 
 
 def main():
@@ -11,27 +22,35 @@ def main():
     if not os.path.exists(logger_file_name):
         print('Master logger module not found')
         return -1
-    master_logger_path = os.path.abspath(logger_file_name)
+    master_logger_path = os.path.realpath(logger_file_name)
     print(master_logger_path)
     dirs = list(filter(os.path.isdir, os.listdir('.')))
     while dirs:
         elem = dirs.pop()
-        c_dir = os.path.abspath(elem)
-        dirs.extend(list(filter(os.path.isdir, os.listdir(c_dir))))
         if 'logger.py' in os.listdir(elem):
-            elem_path = os.path.abspath(elem)
+            elem_path = os.path.realpath(elem)
             child_logger_path = os.path.join(elem_path, logger_file_name)
-            try:
-                os.unlink(child_logger_path)
-            except:
-                print('Logger in {} delete failed'.format(elem_path))
-                continue
+            if md5_file(master_logger_path) != md5_file(child_logger_path):
+                try:
+                    os.unlink(child_logger_path)
+                except:
+                    print('Logger in {} delete failed'.format(elem_path))
+                    continue
+                else:
+                    print('Logger in {} deleted, replacing'.format(elem_path))
+                    with open(master_logger_path, 'r') as master:
+                        try:
+                            child = open(child_logger_path, 'w')
+                        except PermissionError as e:
+                            print(e)
+                            continue
+                        else:
+                            with child:
+                                child.write(master.read())
+                    print('Logger in {} replaced'.format(elem_path))
             else:
-                print('Logger in {} deleted, replacing'.format(elem_path))
-                with open(master_logger_path, 'r') as master, \
-                        open(child_logger_path, 'w') as child:
-                    child.write(master.read())
-                print('Logger in {} replaced'.format(elem_path))
+                print('Logger in {} matches the master'
+                      .format(child_logger_path))
 
 if __name__ == '__main__':
     main()
